requirements.txt
requests
beautifulsoup4
python-dotenv
fpdf2
pytz
